Tutorial's episode 18
My take on the tutorial's episode 18. In this episode, we've created a single texture file for all block types. I've turned the TIleMap object into a child class of Manager, since it deals with multiple entities. This allows to control the tile objects (now simply entities) more easily. Originally, I wanted to make the Tiles' textures point to a single one in the TileMap object, but it proved difficult because of how C++ works.